Bubble Sort function
Created a file with a bubble sort function definition. This function
repeatedly steps through the input list element by element, comparing
the current element with the one after it, swapping their values if
necessary by invoking the exchange function defined in the same file

The exchange function is passed references to the array elements to
swap; it then swaps the values with the aid of a temporary variable. It
returns no value.

Finally, updated the header file with the prototype of the bubble
sort function. The program executes without error and complies with betty
style compliance checks